+++ library/ruler/src/main/java/io/mehow/ruler/ImperialLengthUnit.kt
@@ -1,7 +1,6 @@
 package io.mehow.ruler
 
 import java.math.BigDecimal
-import java.math.RoundingMode.DOWN
 import kotlin.Double.Companion.MAX_VALUE
 
 enum class ImperialLengthUnit(
@@ -28,10 +27,6 @@ enum class ImperialLengthUnit(
       return meters >= super.applicableRange.start
     }
   };
-
-  override fun toMeasuredLength(meters: BigDecimal): BigDecimal {
-    return meters.divide(meterRatio, 9, DOWN)
-  }
 
   override fun appliesRangeTo(meters: BigDecimal): Boolean {
     return meters >= applicableRange.start && meters < applicableRange.endInclusive