+++ docs/changelog.md
@@ -13,10 +13,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `io.mehow.ruler:android` to `io.mehow.ruler:ruler-android` artifact.
 - Changelog format follows now [Keep a Changelog](https://keepachangelog.com/) format. Format is applied retroactively to this file.
 - Upgrade to Kotlin `1.4.10`.
+- `Length` `measureLength` property to `measure`.
 
 ### Removed
 - `toDistance()` methods from `LengthUnit` interface.
 - `-Xjvm-default=enable` compiler argument is no longer required.
+- `toMeasuredLength()` function from `LengthUnit` interface.
 
 ## [0.5.3] - 2020-06-14